restructure package & add checkpoints
